\name{examples.AS}
\alias{examples.AS}
\title{Abreu & Sannikov's arbitrary game example}
\description{Example model definition function. Records an arbitrary game with three actions for each player and returns the a list of actions, payoffs, and a discount factor to define the game.  These can then be used by \code{model.initiate} to generate a model description used in the body of the AS algorithm.  This example also contains answers in the 'ans' list entry generated by Richard Kratzwer's rgsolve algorithm, when the default model options are passed.  These are used in accuracy tests.}
\usage{examples.AS(opts)}
\seealso{
  \code{\link{model.initiate}} \code{\link{examples.PD}}, \code{\link{examples.cournot}}, \code{\link{examples.sexes}}, \code{\link{examples.FL.union}}
}
\examples{
model.defn <- examples.AS()
model.defn$delta 
  # Print the discount factor
model <- model.initiate( examples.AS )
  # Pre-process the model
sol <- abSan.eqm( model=model )
  # Solve
}