\name{examples.FL.union}
\alias{examples.FL.union}
\title{Fuchs & Lippi monetary union example}
\description{Example model definition function. Records a version of Fuchs & Lippi's independent monetary policy game.  Here, monetary policy "spills over" from one country to another, so policymakers target not only their own inflation rate, but also it relative to their neightbor's.  This function returns the a list of actions, payoffs, and a discount factor to define the game.  These can then be used by \code{model.initiate} to generate a model description used in the body of the AS algorithm.  This example also contains answers in the 'ans' list entry generated by Richard Kratzwer's rgsolve algorithm, when the default model options are passed.  These are used in accuracy tests.}
\usage{examples.FL.union(opts)}
\arguments{
The list \code{opts} should be passed with the following entries:
  \item{iActs}{is the integer number of actions available to each player. Default is 40.}
  \item{piRange}{the range f inflation choices for each country.  Default is (-5,5).}
}
\seealso{
  \code{\link{model.initiate}} \code{\link{examples.PD}}, \code{\link{examples.sexes}}, \code{\link{examples.AS}}, \code{\link{examples.cournot}}
}
\examples{
model.defn <- examples.FL.union()
model.defn$delta 
  # Print the discount factor
model <- model.initiate( examples.FL.union )
  # Pre-process the model
sol <- abSan.eqm( model=model )
  # Solve
model <- model.initiate( examples.FL.union, opts=list( 'iActs' = 80, piRange=c(-8,8) ) )
sol <- abSan.eqm( model=model )
  # Alternative use
}