\name{examples.sexes}
\alias{examples.sexes}
\title{Battle of the sexes example}
\description{Example model definition function. Records a "Battle of the sexes" game and returns the a list of actions, payoffs, and a discount factor to define the game.  These can then be used by \code{model.initiate} to generate a model description used in the body of the AS algorithm.  This example also contains answers in the 'ans' list entry generated by Richard Kratzwer's rgsolve algorithm.  These are used in accuracy tests.}
\usage{examples.sexes(opts)}
\seealso{
  \code{\link{model.initiate}} \code{\link{examples.PD}}, \code{\link{examples.cournot}}, \code{\link{examples.AS}}, \code{\link{examples.FL.union}}
}
\examples{
model.defn <- examples.sexes()
model.defn$delta 
  # Print the discount factor
model <- model.initiate( examples.sexes )
  # Pre-process the model
sol <- abSan.eqm( model=model )
  # Solve
}